Prepares to merge VMM screen into graphics module
