+++ gui/src/vmm/cpu/x86_64.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 use super::{CpuManager, GdbError};
+use crate::graphics::Screen;
 use crate::hv::Hypervisor;
-use crate::screen::Screen;
 use gdbstub::target::ext::base::BaseOps;
 use gdbstub::target::ext::breakpoints::{
     Breakpoints, BreakpointsOps, SwBreakpoint, SwBreakpointOps,
